fstat, lseek sistem cagrilari eklendi. Dosya sistemi cagrilarinda duzenlemeler

+++ include/types.h
@@ -22,8 +22,8 @@ typedef uint32_t addr_t;
 // FIXME: newlib types.h
 #ifndef _SYS_TYPES_H
 typedef int key_t;
-typedef int ino_t;
+typedef short ino_t;
 typedef int off_t;
-typedef int dev_t;
+typedef short dev_t;
 #endif
 #endif /* TYPEDEFS_H */
